edited CellTestCase

+++ shoze/core/cell.py
@@ -1,4 +1,5 @@
 from typing import Dict, List, Optional, cast
+import warnings
 from shoze.utils.types import Distances
 
 
@@ -12,7 +13,7 @@ class Cell:
         return self._column
 
     @property
-    def links(self) -> "Cell":
+    def links(self):
         return self._links.keys()
 
     @property
@@ -44,31 +45,36 @@ class Cell:
 
         self._links: Dict[Cell, bool] = {}
 
-    def link(self, cell: "Cell", bidirectionally: bool = True):
+    def link(self, cell: "Cell", bidirectionally: bool = True) -> None:
+        if not isinstance(cell, Cell):
+            raise ValueError("Link can only be made between two cells")
         self._links[cell] = True
         if bidirectionally:
             cell.link(self, False)
 
-    def unlink(self, cell: "Cell", bidirectionally: bool = True):
-        del self._links[cell]
-        if bidirectionally:
-            cell.unlink(self, False)
+    def unlink(self, cell: "Cell", bidirectionally: bool = True) -> None:
+        if not isinstance(cell, Cell):
+            raise ValueError("Unlink can only be made between two cells")
+        elif not self.is_linked(cell):
+            warnings.warn(
+                f"Cell({self.row},{self.column}) and Cell({cell.row},{cell.column}) are not linked"
+            )
+        else:
+            del self._links[cell]
+            if bidirectionally:
+                cell.unlink(self, False)
 
-    def is_linked(self, cell: "Cell"):
+    def is_linked(self, cell: "Cell") -> bool:
         return cell in self._links
 
-    def __eq__(self, cell: "Cell"):
+    def __eq__(self, cell: "Cell") -> bool:
         return self.row == cell.row and self.column == cell.column
 
-    def __ne__(self, cell: "Cell"):
+    def __ne__(self, cell: "Cell") -> bool:
         return not (self.row == cell.row and self.column == cell.column)
 
     def __hash__(self) -> int:
         return hash((self.row, self.column))
 
-    def __repr__(self):
-        return f"Cell {self.row},{self.column}: \n"
-
-
-def is_cell(cell: "Cell") -> bool:
-    return isinstance(cell, Cell)
+    def __repr__(self) -> str:
+        return f"Cell({self.row},{self.column})"
